requirements.txt
Django>=6.0.4
djangorestframework>=3.17.1
djangorestframework-simplejwt==5.3.1
django-unfold>=0.90.0
django-cors-headers==4.3.1
django-filter>=25.2
Pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dateutil>=2.9.0
drf-spectacular==0.27.2
